Move to under passwordPages folder, change classes file directory path, put set error after catch
